refactor(test) put sprockets helpers in module
