+++ test/react/jsx/jsx_prepocessor_test.rb
@@ -1,13 +1,13 @@
 require 'test_helper'
 
-when_sprockets_available do
+SprocketsHelpers.when_available do
   class JSXPreprocessorTest < ActiveSupport::TestCase
     REQUIRED_JAVASCRIPT = "var requirePlainJavascript = true;"
     REQUIRED_COFFEESCRIPT =  "var requireCoffee; requireCoffee = true;"
     REQUIRED_JSX = "React.createElement(\"div\", { className: \"require-jsx\" });"
     OWN_JSX = "React.createElement(\"div\", { className: \"le-javascript\" });"
     test 'executes //= require directives' do
-      require_parent = fetch_asset_body("require_test/jsx_preprocessor_test.js")
+      require_parent = SprocketsHelpers.fetch_asset_body("require_test/jsx_preprocessor_test.js")
 
       assert_compiled_javascript_includes(require_parent, REQUIRED_JAVASCRIPT)
       assert_compiled_javascript_includes(require_parent, REQUIRED_COFFEESCRIPT)
