+++ test/react/jsx/jsx_transformer_test.rb
@@ -1,16 +1,16 @@
 require 'test_helper'
 
-when_sprockets_available do
+SprocketsHelpers.when_available do
   class JSXTransformerTest < ActionDispatch::IntegrationTest
     setup do
       reset_transformer
       React::JSX.transformer_class = React::JSX::JSXTransformer
-      manually_expire_asset('JSXTransformer.js')
+      SprocketsHelpers.manually_expire_asset('JSXTransformer.js')
     end
 
     teardown do
       reset_transformer
-      manually_expire_asset('JSXTransformer.js')
+      SprocketsHelpers.manually_expire_asset('JSXTransformer.js')
     end
 
     test 'can use dropped-in version of JSX transformer' do
@@ -18,7 +18,7 @@ when_sprockets_available do
       replacing_path =  Rails.root.join("vendor/assets/react/JSXTransformer.js")
 
       FileUtils.cp hidden_path, replacing_path
-      manually_expire_asset('example3.js')
+      SprocketsHelpers.manually_expire_asset('example3.js')
 
       get '/assets/example3.js'
       FileUtils.rm replacing_path
@@ -53,7 +53,7 @@ when_sprockets_available do
 
       FileUtils.mkdir_p(custom_path)
       FileUtils.cp(hidden_path, replacing_path)
-      manually_expire_asset('example3.js')
+      SprocketsHelpers.manually_expire_asset('example3.js')
       get '/assets/example3.js'
 
       FileUtils.rm_rf custom_path
