+++ test/react/server_rendering/manifest_container_test.rb
@@ -9,13 +9,13 @@ if defined?(Sprockets::Rails::VERSION) &&
 
   class ManifestContainerTest < ActiveSupport::TestCase
     def setup
-      precompile_assets
+      SprocketsHelpers.precompile_assets
 
       @manifest_container = React::ServerRendering::ManifestContainer.new
     end
 
     def teardown
-      clear_precompiled_assets
+      SprocketsHelpers.clear_precompiled_assets
     end
 
     def test_find_asset_gets_asset_contents
