+++ test/react/server_rendering/yaml_manifest_container_test.rb
@@ -3,13 +3,13 @@ require "test_helper"
 if Rails::VERSION::MAJOR == 3
   class YamlManifestContainerTest < ActiveSupport::TestCase
     def setup
-      precompile_assets
+      SprocketsHelpers.precompile_assets
 
       @manifest_container = React::ServerRendering::YamlManifestContainer.new
     end
 
     def teardown
-      clear_precompiled_assets
+      SprocketsHelpers.clear_precompiled_assets
     end
 
     def test_find_asset_gets_asset_contents
