+++ test/test_helper.rb
@@ -3,6 +3,11 @@ if RUBY_PLATFORM != "java"
   SimpleCov.start
 end
 
+support_path = File.expand_path("../support/*.rb", __FILE__)
+Dir.glob(support_path).each do |f|
+  require(f)
+end
+
 # Configure Rails Environment
 ENV["RAILS_ENV"] = "test"
 
@@ -15,10 +20,7 @@ require "capybara/rails"
 require "capybara/poltergeist"
 Dummy::Application.load_tasks
 
-support_path = File.expand_path("../support/*.rb", __FILE__)
-Dir.glob(support_path).each do |f|
-  require(f)
-end
+
 
 WebpackerHelpers.clear_webpacker_packs
 
@@ -41,65 +43,11 @@ CACHE_PATH = Pathname.new File.expand_path("../dummy/tmp/cache", __FILE__)
 
 Rails.backtrace_cleaner.remove_silencers!
 
-def clear_sprockets_cache
-  # Remove cached files
-  Rails.root.join('tmp/cache').tap do |tmp|
-    tmp.rmtree if tmp.exist?
-    tmp.mkpath
-  end
-end
-
 def reset_transformer
-  clear_sprockets_cache
+  SprocketsHelpers.clear_sprockets_cache
   React::JSX.transformer_class = React::JSX::DEFAULT_TRANSFORMER
   React::JSX.transform_options = {}
   React::JSX.transformer = nil
-end
-
-# Sprockets 2 doesn't expire this assets well in
-# this kind of setting,
-# so override `fresh?` to mark it as expired.
-def manually_expire_asset(asset_name)
-  asset = Rails.application.assets[asset_name]
-  def asset.fresh?(env); false; end
-end
-
-def precompile_assets
-  capture_io do
-    # Changing directories is required because:
-    # - assets:precompile runs webpacker:compile when availabled
-    # - webpacker:compile depends on `./bin/webpack`, so `.` must be the app root
-    Dir.chdir("./test/dummy") do
-
-      ENV['RAILS_GROUPS'] = 'assets' # required for Rails 3.2
-      Rake::Task['assets:precompile'].reenable
-
-      if Rails::VERSION::MAJOR == 3
-        Rake::Task['assets:precompile:all'].reenable
-        Rake::Task['assets:precompile:primary'].reenable
-        Rake::Task['assets:precompile:nondigest'].reenable
-      end
-
-      Rake::Task['assets:precompile'].invoke
-    end
-  end
-
-  if Rails.application.respond_to?(:assets_manifest)
-    # Make a new manifest since assets weren't compiled before
-    config = Rails.application.config
-    path = File.join(config.paths['public'].first, config.assets.prefix)
-    new_manifest = Sprockets::Manifest.new(Rails.application.assets, path)
-    Rails.application.assets_manifest = new_manifest
-  end
-
-  assets_directory = File.expand_path("../dummy/public/assets", __FILE__)
-  raise "Asset precompilation failed" unless Dir.exists?(assets_directory)
-end
-
-def clear_precompiled_assets
-  assets_directory = File.expand_path("../dummy/public/assets", __FILE__)
-  FileUtils.rm_r(assets_directory)
-  ENV.delete('RAILS_GROUPS')
 end
 
 # Rails 3.2's version of MiniTest does not have `capture_io` defined. For
@@ -134,18 +82,6 @@ end
 
 def wait_for_turbolinks_to_be_available
   sleep(1)
-end
-
-
-# The block depends on sprockets, don't run it if sprockets is missing
-def when_sprockets_available
-  if !SKIP_SPROCKETS
-    yield
-  end
-end
-
-def fetch_asset_body(asset_logical_path)
-  Rails.application.assets[asset_logical_path].to_s
 end
 
 # Different processors may generate slightly different outputs,
